+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+- shape.Card default width is 310px
+- shape.Card takes optional note and text
+
 ## [0.27.0] 2023-08-20
 
 - Add package draw/goviz for generating frame sequence diagrams